fix(lsp monaco): avoid runtime and bundling errors on Monaco and LSP
