+++ packages/lsp/src/index.browser.ts
@@ -1,0 +1,2 @@
+export * from '@coderline/alphatab-language-server/index';
+export { startWebWorkerLanguageServer } from '@coderline/alphatab-language-server/server/browser';
