+++ packages/monaco/src/worker.ts
@@ -1,4 +1,4 @@
-import { startWebWorkerLanguageServer } from '@coderline/alphatab-language-server/index';
+import { startWebWorkerLanguageServer } from '@coderline/alphatab-language-server/server/browser';
 
 const workerGlobalThis = globalThis as unknown as DedicatedWorkerGlobalScope;
 startWebWorkerLanguageServer(workerGlobalThis, workerGlobalThis);
